tdf#125445 Add methods for handles in gear6 and gear9 shapes
The shapes 'gear6' and 'gear9' need very special methods to
calculate the adjustment value from the handle position. Inserting
these methods shifts the other cases one level down, therefore the
patch looks large. The gear6 and gear9 shapes do not fit to the
common test pattern, so they are not included in the unit test.

Change-Id: I809e92d87118303332be87512062c431f235d859
Reviewed-on: https://gerrit.libreoffice.org/72796
Tested-by: Jenkins
Reviewed-by: Regina Henschel <rb.henschel@t-online.de>