tdf#124752: We do need to accept text/html from the system clipboard
Sadly we can't accept public.utf8-plain-text, i.e.
text/plain;charset=utf-8, as the code does not know of such a
SotClipboardFormatId. The only text/plain is UTF-16.

And it probably is a good idea to accept HTML anyway. We have to
continue to (I think) not offer text/html because of the assertion
failure issue, see comment in DataFlavorMapper::openOfficeToSystemFlavor().

Change-Id: If5d77b97649424e347c50af10475c2be997c8632