Resolves: tdf#123752 allow group separator in formula values if possible
If possible here means if it can't be an operator (is not an ASCII
value) and is not any other separator. This restores the "working
by chance" behaviour for some locales that use NO-BREAK SPACE as
group separator. It never worked for locales that use one of the
"ordinary" comma or dot or apostrophe group separator and will
not.

Change-Id: I8b7ba1b502b6b7367ffb2199d2f3922ab605a659
Reviewed-on: https://gerrit.libreoffice.org/73422
Reviewed-by: Eike Rathke <erack@redhat.com>
Tested-by: Jenkins