tdf#121486: Correct scale of broken fonts on Windows
Some fonts have negative tmAveCharWidth which makes no sense (fonts can't have
negative glyph width, so the average can't be negative) and it would cause our
code to apply negative scale to the glyphs of these fonts making them drawn on
the flipped horizontally.

Fix this by using the absolute value so it is always positive.

Change-Id: I731e7aad80dae734847679a1b08c6ac78111e16c
Reviewed-on: https://gerrit.libreoffice.org/74109
Tested-by: Jenkins
Reviewed-by: Khaled Hosny <khaledhosny@eglug.org>