Fix INetURLObject::setName description
...after 359e0b47a0f96ffa595a0c38a5e5318d797812fe "loplugin:unuseddefaultparams"
had removed its nIndex and bIgnoreFinalSlash parameters denoting the segment
whose name to set.

(One would assume that a hierarchical URL always has a last segment, so the
"[if] a last segment does not exist, false is returned" part be removed from the
description.  But for a hierarchical URL with an empty (i.e., just "/" in the
canonic form) path, setName("foo") actually returns false.)

Change-Id: I35b96f4687784648c141e16c0806fe49d80c940a
Reviewed-on: https://gerrit.libreoffice.org/74487
Tested-by: Jenkins
Reviewed-by: Stephan Bergmann <sbergman@redhat.com>