tdf#120873 Qt5 invert checked state on trigger
SalMenu's event handling is not really compatible with QAction.
LO expects to get something like an activate event, which happens
without any menu state update. The item handler will then update
the item state, as expected by LO.

I'm not sure how this could be implemented in a good way in Qt, so
this patch is just a hack. If the item is checkable, we invert its
state before handling it. This just works correctly for single items,
not radio groups, but it does the trick for this bug. And since
setChecked() just emits toggled(), triggered() won't run twice.

Maybe someone can come up with a better way using event filters, if
ever needed.

Change-Id: Ic32d07525ce8f394b83d04f53689d1669ddc9d83
Reviewed-on: https://gerrit.libreoffice.org/74590
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Reviewed-by: Jan-Marek Glogowski <glogow@fbihome.de>