tdf#125706 and tdf#125665 writer, speed up RDF
The RDF stuff is sloooooooow, so (a) add some caching and (b) fold a
very hot UNO_QUERY call.

To add the caching we need to add a new UNO interface, since
XEnumeration is not amenable to being cached.

Add an optimised getStatementsGraph_NoLock2 that skips the intermediate
enumeration object, and consequently a lot of locking/unlocking.

Cache by OUString key, to avoid expensive OUString<->OString conversion
when looking up entries in the cache.

For the test document in tdf#125706, this takes the time from 7s to 5s for me.

For the test document in tdf#125665, this takes the load time
from 60s to 7s for me.

Change-Id: I207387e975b4f107834edd0974134c481fb4012d
Reviewed-on: https://gerrit.libreoffice.org/74740
Tested-by: Jenkins
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>