tdf#125554 PPTX export: handle gradient transparency for gradient fill
Regression from commit cfc1f4ea4889f768d689a0df71519e9bcb707bc0 (oox:
disable gradient fill grab-bag for PPTX, 2019-02-05), the problem was
that in the past grab-bag roundtrip worked (in some cases) for this
shape fill case, but true roundtrip did not.

So when the commit disabled grab-bags (since their color pointers in the
theme don't work in the PPTX case), a previously not implemented feature
now started causing a real problem.

Fix the bug by adding support for transparent linear gradients on the
exports side. This means that in case the import creates both a fill
gradient and a transparency gradient, then now the export creates markup
based on both, not only based on the fill gradient.

Change-Id: I99fa3caba2b2884c2acb7e0704bbeb0b6cffd4a4
Reviewed-on: https://gerrit.libreoffice.org/74968
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>