vcl: make it possible to select specific scaling algorithm
Default, Fast, BestQuality scaling flags are used for selecting
the best sclaing algorithm for a specific task, but not all
specialized sclaing algorithms have its own flag (Super,
NearestNeighbor) and are just selectable using one of the above.
This adds the missing flags so it's possible to select a specific
algorithm.

Change-Id: Ied41f27a21a4fcc799537396f9077a9c77cc1c60
Reviewed-on: https://gerrit.libreoffice.org/75759
Reviewed-by: Chris Sherlock <chris.sherlock79@gmail.com>
Tested-by: Jenkins
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>