tdf#125863 Rework tdf#107593 so header footer delete doesn't crash
Header footer delete from the last page when last page is unneeded
(stack level 24 SwRootFrame::RemoveSuperfluous /// remove pages that are
not needed at all) causes SwFrameMenuButtonBase::dispose() which sets
m_pFrame to nullptr which causes crash when IsEmptyHeaderFooter calls
GetPageFrame(). Doing ToggleHeaderFooterEdit() when header or footer
deletion is accepted in SwWrtShell::ChangeHeaderOrFooter avoids the
crash. SwHeaderFooterWin is also disposed which would make the
GrabFocusToDocument crash if it got that far so use the shell window
which is still valid here to GrabFocusToDocument.

Change-Id: Iaced23e6716442671c5ea29b79be178f4fc836d3
Reviewed-on: https://gerrit.libreoffice.org/75010
Reviewed-by: Michael Stahl <Michael.Stahl@cib.de>
Tested-by: Michael Stahl <Michael.Stahl@cib.de>