sw: remove dynamic_cast to MarkBase in MarkManager
Forward port of https://gerrit.libreoffice.org/#/c/74851/
and much smaller because the shared_ptr has been removed in commit
9ff648c691f003a11eba9a22ac37032d72b4b642

The MarkManager always has MarkBase instances in its vectors, so
abstract the vector better in the IDocumentMarkAccess interface by
exposing only a wrapper iterator that converts from MarkBase to
IMark.

There is an unexpected pitfall here in that forward iterator's
operator* is required to return a reference, but at least for a
const_iterator it appears to work to just return a value instead.

The dynamic_cast of every element in some MarkManager functions is
quite expensive and shows as 10-15% in callgrind.

In bubli's measurements, in the 6.1 branch this speeds up loading the
bugdoc by ~40s of the initial 3min 50s.

Change-Id: I65d974bea0301bf7b4cfa25ad6daae10b6768202
Reviewed-on: https://gerrit.libreoffice.org/75033
Tested-by: Jenkins
Reviewed-by: Michael Stahl <Michael.Stahl@cib.de>