LOK: Emit the doc. modified state even when changing the cell inline.
The "LOK: emit modified status on entering cell data" patch did this
only for the formula input line, and did not change the state when the
user just clicked on the cell and started modifying it; so this patch
extends it.

Also change the actual DocumentModified state of the document directly
instead of just sending a flag, so that the state in LOOLWSD and in the
core don't get out of sync.

Change-Id: Ie516ce5a74d94247d79cf993517d7f4a6657cc94
Reviewed-on: https://gerrit.libreoffice.org/75701
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Jan Holesovsky <kendy@collabora.com>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/95516
Tested-by: Jenkins
Reviewed-by: Michael Meeks <michael.meeks@collabora.com>