tdf#126597 don't use modeless dialog for senddoc.exe
There is already a large comment in initAttachmentList about
problems with MAPI_DIALOG_MODELESS. Using MAPI_DIALOG_MODELESS with
Outlook 2016 and multiple attachments sometimes produces a
MAPI_E_FAILURE, but most times it crashes for me. And it's not a
problem with the removed temporary files, as uncommenting that
shows the same problem. And there seems to be many more problems:

https://social.msdn.microsoft.com/Forums/en-US/5d8fece6-9d93-490c-9331-625c17e3291d/mapisendmailhelper-and-mapidialogmodeless

But actually I don't see any blocking of LO, if I switch to
using MAPI_DIALOG, as senddoc.exe already runs in the background.

So this switches MAPI_DIALOG_MODELESS to MAPI_DIALOG.

This reverts commit 5874c76371562c3e2d8564b1fb04df1997091d27
("tdf#116074: Use modeless dialogs with supporting mailers").

Change-Id: Ie0f8f22196d1a174dfeada2bc4aabb1717ee16a7
Reviewed-on: https://gerrit.libreoffice.org/76155
Tested-by: Jenkins
Reviewed-by: Jan-Marek Glogowski <glogow@fbihome.de>