tdf#118856 vcl: avoid flicker in the about dialog
There were two problems here:

1) The flickering widget in question is VclMultiLineEdit, but it wasn't
like its own painting caused a flicker, rather the calling
vcl::Window::PushPaintHelper() already erased the widget and later the
widget painted itself. So the easy way with vcl::BufferDevice is not
enough here. VclMultiLineEdit seems to be working with double buffering
just fine, so enable that in the vcl::AboutDialog ctor.

2) Now that text did not flicker, its background was wrong. In case the
NWF determines some gray background, the background was white for most
text when there was no selection.

When there was selection, then the after-selection area (in a single
line) was fine, but nothing else.

It seems TextEngine::ImpPaint() assumes that no need to set a default
text fill color before drawing text, but the above shows that there is
difference between the starting state and after clearing away a custom
color with SetTextFillColor(). So fix this problem by always explicitly
calling SetTextFillColor() before text paint: either setting the
selection color or the default one.

This brings improvement for the Windows GDI backend; Windows GL, Linux
gen/gtk/gtk3 is unchanged in practice.

Change-Id: Ie7ff3a7f36bedf73424ea579a00588bd8858d6c4
Reviewed-on: https://gerrit.libreoffice.org/77219
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins