tdf#126784 sw: only distribute across selected columns
partial revert of LO6.2 commit ab18c17d70e1dcf5cf9db38256d35e6af479373e
because MERGED cells had a very unexpected effect on the
columns that were altered. Unselected columns are INCLUDED
in the calculation IF they extend over the selected columns
when SwTableSearchType::Col uses the entire table column.

The last table in the unit test demonstrates why the regressive behaviour
was included. A user can make an ugly layout if they don't choose a
good set of columns.

However, the merged cell really changes the way the calculation
works, and since giving power to the user is ultimately the best,
just revert the portion that tries to save the user from themself.

I struggled then and now to get a good enough understanding
to improve the documentation - which really is needed
since reading it didn't help me a lot.
However, since my understanding was and is obviously flawed,
I'll just revert back to the original wording.

Change-Id: Icf8f01f095652e2a023e0398852f43aa6b90332f
Reviewed-on: https://gerrit.libreoffice.org/77976
Tested-by: Jenkins
Reviewed-by: Justin Luth <justin_luth@sil.org>