tdf#98130 vcl CheckBox: fix incorrect ignored mouse event
Regression from commit 53ef550ac6fc3505914c3d77e9d0575913e54bb3 (vcl
button: Delegate RenderContext to sub-methods, use Invalidate,
2015-04-28), the problem was that when clicking on the middle checkbox
in the second row of the bugdoc, the click was simply ignored, the
checkbox widget remained in unchecked state.

The above commit changed CheckBox::GetFocus() so that it uses
Invalidate() instead of ImplDrawCheckBox(), which looks innocent, but
ImplDrawCheckBox() resulted in a call to CheckBox::ImplDraw(), which (as
a side effect) initialized maMouseRect to a non-empty rectangle.

Now in case GetFocus() was followed by MouseButtonDown() immediately,
then that ignored the mouse event as the event position wasn't in the
mouse rectangle.

Fix this by updating the window right after the invalidate, this way we
still don't paint directly, but MouseButtonDown() already has a correct
maMouseRect.

Change-Id: I0371a2d86a44023042cf742dec70d4558ff5de80
Reviewed-on: https://gerrit.libreoffice.org/78021
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins